feat: remove modal close button focus boxShadow
